printer-api/requirements.txt
kafka-python==2.0.2
bambulabs_api
python-dotenv==1.0.0
requests==2.31.0
